+++ executor/executor_legacy_test.go
@@ -108,7 +108,6 @@ var _ = SerialSuites(&testSplitTable{&baseTestSuite{}})
 var _ = SerialSuites(&testSerialSuite1{&baseTestSuite{}})
 var _ = SerialSuites(&globalIndexSuite{&baseTestSuite{}})
 var _ = SerialSuites(&testSerialSuite{&baseTestSuite{}})
-var _ = SerialSuites(&testCoprCache{})
 
 type testSuite struct{ *baseTestSuite }
 type testSuiteP1 struct{ *baseTestSuite }
@@ -116,11 +115,6 @@ type testSuiteP2 struct{ *baseTestSuite }
 type testSplitTable struct{ *baseTestSuite }
 type globalIndexSuite struct{ *baseTestSuite }
 type testSerialSuite struct{ *baseTestSuite }
-type testCoprCache struct {
-	store kv.Storage
-	dom   *domain.Domain
-	cls   testutils.Cluster
-}
 
 // MockGC is used to make GC work in the test environment.
 func MockGC(tk *testkit.TestKit) (string, string, string, func()) {
@@ -4969,77 +4963,6 @@ func (s *testSuite) TestIssue13758(c *C) {
 		"4",
 		"<nil>",
 	))
-}
-
-func (s *testCoprCache) SetUpSuite(c *C) {
-	originConfig := config.GetGlobalConfig()
-	config.StoreGlobalConfig(config.NewConfig())
-	defer config.StoreGlobalConfig(originConfig)
-	cli := &regionProperityClient{}
-	hijackClient := func(c tikv.Client) tikv.Client {
-		cli.Client = c
-		return cli
-	}
-	var err error
-	s.store, err = mockstore.NewMockStore(
-		mockstore.WithClusterInspector(func(c testutils.Cluster) {
-			mockstore.BootstrapWithSingleStore(c)
-			s.cls = c
-		}),
-		mockstore.WithClientHijacker(hijackClient),
-	)
-	c.Assert(err, IsNil)
-	s.dom, err = session.BootstrapSession(s.store)
-	c.Assert(err, IsNil)
-}
-
-func (s *testCoprCache) TearDownSuite(c *C) {
-	s.dom.Close()
-	s.store.Close()
-}
-
-func (s *testCoprCache) TestIntegrationCopCache(c *C) {
-	originConfig := config.GetGlobalConfig()
-	config.StoreGlobalConfig(config.NewConfig())
-	defer config.StoreGlobalConfig(originConfig)
-
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t (a int primary key)")
-	tblInfo, err := s.dom.InfoSchema().TableByName(model.NewCIStr("test"), model.NewCIStr("t"))
-	c.Assert(err, IsNil)
-	tid := tblInfo.Meta().ID
-	tk.MustExec(`insert into t values(1),(2),(3),(4),(5),(6),(7),(8),(9),(10),(11),(12)`)
-	tableStart := tablecodec.GenTableRecordPrefix(tid)
-	s.cls.SplitKeys(tableStart, tableStart.PrefixNext(), 6)
-
-	c.Assert(failpoint.Enable("github.com/pingcap/tidb/store/mockstore/unistore/cophandler/mockCopCacheInUnistore", `return(123)`), IsNil)
-	defer func() {
-		c.Assert(failpoint.Disable("github.com/pingcap/tidb/store/mockstore/unistore/cophandler/mockCopCacheInUnistore"), IsNil)
-	}()
-
-	rows := tk.MustQuery("explain analyze select * from t where t.a < 10").Rows()
-	c.Assert(rows[0][2], Equals, "9")
-	c.Assert(strings.Contains(rows[0][5].(string), "cop_task: {num: 5"), Equals, true)
-	c.Assert(strings.Contains(rows[0][5].(string), "copr_cache_hit_ratio: 0.00"), Equals, true)
-
-	rows = tk.MustQuery("explain analyze select * from t").Rows()
-	c.Assert(rows[0][2], Equals, "12")
-	c.Assert(strings.Contains(rows[0][5].(string), "cop_task: {num: 6"), Equals, true)
-	hitRatioIdx := strings.Index(rows[0][5].(string), "copr_cache_hit_ratio:") + len("copr_cache_hit_ratio: ")
-	c.Assert(hitRatioIdx >= len("copr_cache_hit_ratio: "), Equals, true)
-	hitRatio, err := strconv.ParseFloat(rows[0][5].(string)[hitRatioIdx:hitRatioIdx+4], 64)
-	c.Assert(err, IsNil)
-	c.Assert(hitRatio > 0, Equals, true)
-
-	// Test for cop cache disabled.
-	cfg := config.NewConfig()
-	cfg.TiKVClient.CoprCache.CapacityMB = 0
-	config.StoreGlobalConfig(cfg)
-	rows = tk.MustQuery("explain analyze select * from t where t.a < 10").Rows()
-	c.Assert(rows[0][2], Equals, "9")
-	c.Assert(strings.Contains(rows[0][5].(string), "copr_cache: disabled"), Equals, true)
 }
 
 func (s *testSerialSuite) TestCoprocessorOOMTicase(c *C) {
